Refactoring frisbee conn into async and stream structs
